Revamp
The previous implementation turned out to be way to slow for the purpose
it was designed for. I gutted it and used a simple count-based cache
limit instead of using sys.getsizeof().  I offloaded the flush operation
to a background thread so the process only waits when it needs to access
the shelf and the shelf is being written to.